+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## 0.12.1 (????-??-??)
 
-## Bug fixes
+### Bug fixes
 
 -   Fix regression in reading date columns (#616)
 -   Fix error in `read_dataframe` when `use_arrow=True` and `columns` is used to filter