+++ CHANGES.md
@@ -1,11 +1,11 @@
 # CHANGELOG
 
-## 0.12.1 (????-??-??)
+## 0.12.1 (2025-11-28)
 
 ### Bug fixes
 
 -   Fix regression in reading date columns (#616)
--   Fix error in `read_dataframe` when `use_arrow=True` and `columns` is used to filter
+-   Fix regression in `read_dataframe` when `use_arrow=True` and `columns` is used to filter
     out columns of some specific types (#611)
 -   Fix Time type columns being skipped when reading with `arrow=False` (#617)